bug fix, AllDead не идеален
